fix(ops): refuse a substituted function body or a trigger that does not stamp
Two ways a restored database could pass adoption while the boundary it names
was not the boundary that runs.

Any `(uuid)` SECURITY DEFINER function with a pinned search_path was accepted
and handed to the provisioner with control-role execution, without a look at its
language, return type, or body. It now has to be plpgsql, return void, and still
reference catalog.tenants, the data_t_ prefix, and the tenant advisory lock.
That is a structural check, not proof of provenance — pinning the migration's
exact text in app code would go stale the next time a migration edits the body,
in a module whose job is to still work years after the restore — and the
docstring says so, along with why the residual exposure only runs one way:
adoption moves these functions from the restoring superuser with PUBLIC EXECUTE
to a NOLOGIN role with none, so it can fail to notice a tampered body but never
raise its privilege.

And the boundary trigger was matched by name and enabled state alone. One
recreated under that name on the wrong event, or pointed at another function,
counted as live. It now has to be BEFORE INSERT FOR EACH ROW on the migration's
stamping function. `stamping_trigger_disabled` becomes `stamping_trigger_inert`
because disabled is no longer the only way to hold the name and stamp nothing.

Also fixes the Docs Contract job: the public-surface gate allows exactly three
occurrences of the tenancy term in RUNBOOK.md and the new prose added a fourth.
Rephrased rather than widening the allowlist.

tenant_adoption.py crossed the 1000-line module ceiling on the way, so the
report types, the success predicate, and the operator-facing text moved to
tenant_adoption_report.py — what adoption found, split from what adoption does.
